flink-streaming-host-network-data-usage/data-source/src/main/resources/requirements.txt
avro==1.8.1
kafka-python==1.3.5

